Add boost dependency to fix Python handling
Before, when the first Python interpreter on the PATH had a different
version than the one we linked, the test executables failed at startup.
boost now has to be available during build time. I tried to download it
automatically when it is not available (as we do for the other
dependencies) but that proved to be very complicated (linking was no
problem, but installing nias_cpp afterwards was a big problem because we
then also have to export all of the boost targets manually etc.)